seo-monitor/requirements.txt
google-api-python-client>=2.0
google-auth>=2.0
google-auth-httplib2>=0.1
requests>=2.28
python-dotenv>=1.0
pysocks>=1.7
